fix update_trade_data to handle open_trade is not exists

+++ recorder.py
@@ -59,22 +59,24 @@ def update_trade_data(table_name):
         + 'where state=\'OPEN\''
         ,conn
     )['tradeId'])
-    #joinするため文字列型に変換
-    open_ids = list(map(str, open_ids))
 
-    fetched_trades = pd.DataFrame(columns=trades_header)
-    for id in open_ids:
-        trade = oanda_api.get_trade(id)
-        s = pd.Series(trade)
-        fetched_trades = fetched_trades.append(s,ignore_index=True)
+    if len(open_ids) > 0:
+        #joinするため文字列型に変換
+        open_ids = list(map(str, open_ids))
 
-    conn.execute(
-        'delete from '+ table_name + ' where tradeId in ('
-        + ','.join(open_ids) + ');'
-    )
-    conn.commit()
+        fetched_trades = pd.DataFrame(columns=trades_header)
+        for id in open_ids:
+            trade = oanda_api.get_trade(id)
+            s = pd.Series(trade)
+            fetched_trades = fetched_trades.append(s,ignore_index=True)
 
-    fetched_trades.to_sql(table_name, conn, if_exists="append", index=False)
+        conn.execute(
+            'delete from '+ table_name + ' where tradeId in ('
+            + ','.join(open_ids) + ');'
+        )
+        conn.commit()
+
+        fetched_trades.to_sql(table_name, conn, if_exists="append", index=False)
 
 def update_price_data():
     params = {
